+++ CHANGELOG.md
@@ -75,3 +75,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 -   Add Feedback components
 -   Add Layout components
 -   Add Mockup components
+
+## [1.0.0] - 2024-20-1
+
+-   Add class and className Switcher
+-   Add astro language support